ref(slam): refactors all around

+++ src/perception_sensor_lib/include/perception_sensor_lib/loop_closure/lap_counter.hpp
@@ -3,7 +3,7 @@
 #include "loop_closure.hpp"
 
 /**
- * @brief Implementation of LoopClosure that detects when the robot returns near 
+ * @brief Implementation of LoopClosure that detects when the robot returns near
  *        the origin and re‑observes any of the first X cones from the map.
  */
 class LapCounter : public LoopClosure {
@@ -26,14 +26,12 @@ public:
    * @param observations     raw observations (unused in this implementation)
    * @return                 {true,0.0} once you re‑see any of map_cones[0..X-1]
    */
-  Result detect(
-    const Eigen::Vector3d& current_pose,
-    const Eigen::VectorXi& map_cones,
-    const Eigen::VectorXi& associations,
-    const Eigen::VectorXd& observations) const override;
+  Result detect(const Eigen::Vector3d& current_pose, const Eigen::VectorXd& map_cones,
+                const Eigen::VectorXi& associations,
+                const Eigen::VectorXd& observations) const override;
 
 private:
-  double threshold_dist_;
+  double threshold_dist_;  // TODO: make this const
   int first_x_cones_;
   int border_width_;
   int minimum_confidence_;
